\name{EMA-package}
\Rdversion{1.1}
\alias{EMA-package}
\alias{EMA}
\docType{package}
\title{
EMA - Easy Microarray Analysis
}
\description{
Numerous analysis methods and tools have been developed to study microarray, many of them being implemented as free R and/or Bioconductor packages.
This abundance of methods makes choosing the best approach difficult for newcomers and non-specialist users.
Based on the experience of the biostatisticians of Institut Curie, we propose a clear analysis strategy combining a large variety of standard
methodologies. The most usual and relevant R functions needed to perform these analyses were selected and gathered in the R package EMA (Easy Microarray data Analysis).
EMA covers an entire analysis process including quality control, normalisation, exploratory analysis,unsupervised and supervised classification, functional analysis and
censored data exploration.
The package can be used for both one or two colours gene expression micrarrays and for exon expression experiments.
}
\details{
\tabular{ll}{
Package: \tab EMA\cr
Type: \tab Package\cr
Version: \tab 1.3.1\cr
Date: \tab 2011-06-09\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
For details, see vignette.
}
\author{
N. Servant, E. Gravier, P. Gestraud, C. Paccard, C. Laurent, I. Brito, J. Mandel, A. Biton, B. Asselain, E. Barillot, P. Hupe.
Maintainer: <pierre.gestraud@curie.fr>
}
\references{
  EMA - A R package for Easy Microarray data Analysis 
}
\keyword{ package }
