\name{htmlheader}
\Rdversion{1.1}
\alias{htmlheader}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{htmlheader}
\description{
Internal function
}
\usage{
htmlheader(towrite, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{towrite}{}
  \item{filename}{}
}


\keyword{internal}
